+++ ProjectWork20-21/static/success.ejs
@@ -97,7 +97,7 @@ div.absolute2{
 
 <body>
 <div class="absolute2">
-<button class="btn2 info" onclick="window.location.href='stulogin.html';">Log Out </button>
+<button class="btn2 info" onclick="window.location.href='/stulogin';">Log Out </button>
 </div>
 </body>
